+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ## Bugfixes
 - Fix transmitted choices in checkbox voting !1847 @alex.simm
+- Fix frontend validation in new poll form !1854 #975 @alex.simm
 
 ## Refactoring
 - Remove unused code from the Flourish library !1789 @\_fridtjof_